chore: Review what has been done

+++ app/src/main/java/com/example/kala/screens/AboutExchangeScreen.kt
@@ -35,6 +35,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
+import com.example.kala.configuration.ABOUT_EXCHANGE_SCREEN_ROUTE
 import com.example.kala.configuration.EDIT_EXCHANGE_SCREEN_ROUTE
 import com.example.kala.configuration.FooterConfiguration
 import com.example.kala.configuration.HeaderConfiguration
@@ -50,9 +51,11 @@ import com.example.kala.screens.components.Footer
 import com.example.kala.screens.components.Header
 import com.example.kala.screens.components.Title
 import com.example.kala.screens.components.buttons.MediumButton
+import com.example.kala.screens.components.popups.ConfirmationPopUp
 import com.example.kala.ui.theme.BoneWhite
-import com.example.kala.ui.theme.Green0
+import com.example.kala.ui.theme.Green1
 import com.example.kala.ui.theme.Red0
+import com.example.kala.ui.theme.dimens
 
 /**
  * Composable function for displaying the About Exchange screen.
@@ -72,7 +75,7 @@ fun AboutExchangeScreen(
 
     val svgFile = MoneyExchangeScope.getSVGFile(moneyExchange.scope)
     val valueSymbol = if (moneyExchange.type == MoneyExchangeType.EXPENSE) "-" else "+"
-    val valueColor = if (moneyExchange.type == MoneyExchangeType.EXPENSE) Red0 else Green0
+    val valueColor = if (moneyExchange.type == MoneyExchangeType.EXPENSE) Red0 else Green1
     val valueText = valueSymbol + moneyExchange.getFormattedValue() + "€" //TODO € should be dynamic
 
     var leftButtonTriggered by remember {
@@ -81,30 +84,48 @@ fun AboutExchangeScreen(
     val onLeftTriggered = {
         leftButtonTriggered = true
     }
-    var rigthButtonTriggered by remember {
+    var rightButtonTriggered by remember {
         mutableStateOf(false)
     }
-    val onRigthTriggered = {
-        rigthButtonTriggered = true
+    val onRightTriggered = {
+        rightButtonTriggered = true
     }
 
-    if (rigthButtonTriggered){
-        rigthButtonTriggered = false
+    if (rightButtonTriggered){
+        rightButtonTriggered = false
         navController?.navigate(route = "$EDIT_EXCHANGE_SCREEN_ROUTE/$monthAssociated/$exchange")
     }
 
     if (leftButtonTriggered){
-        leftButtonTriggered = false
-        MoneyExchangeService.deleteMoneyExchange(monthAssociated, exchange)
+        ConfirmationPopUp(
+            onConfirmButton = {
+                leftButtonTriggered = false
+                MoneyExchangeService.deleteMoneyExchange(monthAssociated, exchange)
+                navController?.navigate(route = RECORD_SCREEN_ROUTE)
+            },
+            onDismissButton = { leftButtonTriggered = false },
+        )
+    }
+
+    val goBackTriggered = {
         navController?.navigate(route = RECORD_SCREEN_ROUTE)
     }
 
     Scaffold(
         topBar = {
-            Header(configuration = HeaderConfiguration.REGISTERED_USER, navController)
+            Header(
+                configuration = HeaderConfiguration.REGISTERED_USER,
+                navController = navController,
+                triggerScreen = ABOUT_EXCHANGE_SCREEN_ROUTE,
+            )
         },
         bottomBar = {
-            Footer(configuration = FooterConfiguration.BACK_AND_HOME, navController)
+            Footer(
+                configuration = FooterConfiguration.BACK_AND_HOME,
+                navController = navController,
+                changeGoBackButton = true,
+                goBackTriggered = goBackTriggered,
+                )
         },
     ){
         Column(
@@ -113,19 +134,19 @@ fun AboutExchangeScreen(
                 .background(BoneWhite),
             horizontalAlignment = Alignment.CenterHorizontally,
         ) {
-            Spacer(modifier = Modifier.padding(50.dp))
+            Spacer(modifier = Modifier.padding(dimens.space4))
             Title(configuration = TitleConfiguration.MORE_INFO)
-            Spacer(modifier = Modifier.padding(10.dp))
+            Spacer(modifier = Modifier.padding(dimens.space0))
             ValueSection(valueText, valueColor)
-            Spacer(modifier = Modifier.padding(10.dp))
+            Spacer(modifier = Modifier.padding(dimens.space0))
             ScopeSection(moneyExchange, svgFile)
-            Spacer(modifier = Modifier.padding(10.dp))
+            Spacer(modifier = Modifier.padding(dimens.space0))
             DateSection(moneyExchange)
-            Spacer(modifier = Modifier.padding(10.dp))
+            Spacer(modifier = Modifier.padding(dimens.space0))
             DescriptionSection(moneyExchange)
-            Spacer(modifier = Modifier.padding(10.dp))
-            OptionSection(onLeftTriggered, onRigthTriggered)
-            Spacer(modifier = Modifier.padding(50.dp))
+            Spacer(modifier = Modifier.padding(dimens.space0))
+            OptionSection(onLeftTriggered, onRightTriggered)
+            Spacer(modifier = Modifier.padding(dimens.space4))
         }
     }
 }
@@ -140,11 +161,11 @@ fun AboutExchangeScreen(
 private fun ValueSection(valueText: String, valueColor: Color) {
     Box(
         modifier = Modifier
-            .width(250.dp)
-            .height(50.dp)
-            .clip(RoundedCornerShape(10.dp))
+            .width(dimens.width6)
+            .height(dimens.height1)
+            .clip(RoundedCornerShape(dimens.rounded))
             .background(Color.White)
-            .border(2.dp, Color.Black, RoundedCornerShape(10.dp))
+            .border(dimens.border, Color.Black, RoundedCornerShape(dimens.rounded))
         ,
         contentAlignment = Alignment.Center
 
@@ -172,11 +193,11 @@ private fun ValueSection(valueText: String, valueColor: Color) {
 private fun ScopeSection(moneyExchange: MoneyExchange, svgFile: Int) {
     Box(
         modifier = Modifier
-            .height(80.dp)
-            .width(250.dp)
-            .clip(RoundedCornerShape(10.dp))
+            .height(dimens.height3)
+            .width(dimens.width6)
+            .clip(RoundedCornerShape(dimens.rounded))
             .background(Color.White)
-            .border(2.dp, Color.Black, RoundedCornerShape(10.dp)),
+            .border(dimens.border, Color.Black, RoundedCornerShape(dimens.rounded)),
         contentAlignment = Alignment.Center
     ) {
         Row(
@@ -186,7 +207,7 @@ private fun ScopeSection(moneyExchange: MoneyExchange, svgFile: Int) {
             Box(
                 modifier = Modifier
                     .fillMaxHeight()
-                    .width(125.dp)
+                    .width(dimens.width2)
                 ,
                 contentAlignment = Alignment.Center
             ){
@@ -201,14 +222,14 @@ private fun ScopeSection(moneyExchange: MoneyExchange, svgFile: Int) {
             Box(
                 modifier = Modifier
                     .fillMaxHeight()
-                    .width(125.dp)
+                    .width(dimens.width2)
                 ,
                 contentAlignment = Alignment.Center
             ){
                 Box(
                     modifier = Modifier
-                        .border(2.dp, Color.Black, shape = CircleShape)
-                        .size(50.dp)
+                        .border(dimens.border, Color.Black, shape = CircleShape)
+                        .size(dimens.image3)
                         .clip(CircleShape)
                         .background(Color.White)
                         .padding(5.dp),
@@ -235,11 +256,11 @@ private fun ScopeSection(moneyExchange: MoneyExchange, svgFile: Int) {
 private fun DateSection(moneyExchange: MoneyExchange) {
     Box(
         modifier = Modifier
-            .height(60.dp)
-            .width(250.dp)
-            .clip(RoundedCornerShape(10.dp))
+            .height(dimens.height2)
+            .width(dimens.width6)
+            .clip(RoundedCornerShape(dimens.rounded))
             .background(Color.White)
-            .border(2.dp, Color.Black, RoundedCornerShape(10.dp))
+            .border(dimens.border, Color.Black, RoundedCornerShape(dimens.rounded))
         ,
         contentAlignment = Alignment.Center
     ) {
@@ -265,11 +286,11 @@ private fun DateSection(moneyExchange: MoneyExchange) {
 private fun DescriptionSection(moneyExchange: MoneyExchange) {
     Box(
         modifier = Modifier
-            .height(150.dp)
-            .width(250.dp)
-            .clip(RoundedCornerShape(10.dp))
+            .height(dimens.height5)
+            .width(dimens.width6)
+            .clip(RoundedCornerShape(dimens.rounded))
             .background(Color.White)
-            .border(2.dp, Color.Black, RoundedCornerShape(10.dp))
+            .border(dimens.border, Color.Black, RoundedCornerShape(dimens.rounded))
             .padding(10.dp)
     ) {
         moneyExchange.description?.let { it1 ->
@@ -290,14 +311,14 @@ private fun DescriptionSection(moneyExchange: MoneyExchange) {
  * Composable function for displaying the option section.
  *
  * @param onLeftTriggered Callback for left button action.
- * @param onRigthTriggered Callback for right button action.
+ * @param onRightTriggered Callback for right button action.
  */
 @Composable
-private fun OptionSection(onLeftTriggered: () -> Unit, onRigthTriggered: () -> Unit) {
+private fun OptionSection(onLeftTriggered: () -> Unit, onRightTriggered: () -> Unit) {
     Row {
         MediumButton(configuration = MediumButtonConfiguration.DELETE, onLeftTriggered)
         Spacer(modifier = Modifier.padding(10.dp))
-        MediumButton(configuration = MediumButtonConfiguration.EDIT, onRigthTriggered)
+        MediumButton(configuration = MediumButtonConfiguration.EDIT, onRightTriggered)
     }
 }
